Refactor: Introduce a higher level unlockCourtCase function
This will be responsible for calling updateLockStatusToUnlocked and storeAuditLogEvents
Renamed the old unlockCourtCase function to updateLockStatusToUnlocked, this is a lower level operation that is called in other transactions
