+++ test/services/resolveCourtCase.integration.test.ts
@@ -11,7 +11,7 @@ import { differenceInMilliseconds } from "date-fns"
 import { ManualResolution, ResolutionReasonCode } from "types/ManualResolution"
 import { TestTrigger, insertTriggers } from "../utils/manageTriggers"
 import insertNotes from "services/insertNotes"
-import unlockCourtCase from "services/unlockCourtCase"
+import updateLockStatusToUnlocked from "services/updateLockStatusToUnlocked"
 import storeAuditLogEvents from "services/storeAuditLogEvents"
 import courtCasesByOrganisationUnitQuery from "services/queries/courtCasesByOrganisationUnitQuery"
 import createAuditLog from "../helpers/createAuditLog"
@@ -21,7 +21,7 @@ import deleteFromDynamoTable from "../utils/deleteFromDynamoTable"
 
 jest.setTimeout(100000)
 jest.mock("services/insertNotes")
-jest.mock("services/unlockCourtCase")
+jest.mock("services/updateLockStatusToUnlocked")
 jest.mock("services/storeAuditLogEvents")
 jest.mock("services/queries/courtCasesByOrganisationUnitQuery")
 
@@ -53,7 +53,9 @@ describe("resolveCourtCase", () => {
     jest.resetAllMocks()
     jest.clearAllMocks()
     ;(insertNotes as jest.Mock).mockImplementation(jest.requireActual("services/insertNotes").default)
-    ;(unlockCourtCase as jest.Mock).mockImplementation(jest.requireActual("services/unlockCourtCase").default)
+    ;(updateLockStatusToUnlocked as jest.Mock).mockImplementation(
+      jest.requireActual("services/updateLockStatusToUnlocked").default
+    )
     ;(storeAuditLogEvents as jest.Mock).mockImplementation(jest.requireActual("services/storeAuditLogEvents").default)
     ;(courtCasesByOrganisationUnitQuery as jest.Mock).mockImplementation(
       jest.requireActual("services/queries/courtCasesByOrganisationUnitQuery").default
@@ -500,7 +502,9 @@ describe("resolveCourtCase", () => {
     })
 
     it("Should return the error if fails to unlock the case", async () => {
-      ;(unlockCourtCase as jest.Mock).mockImplementationOnce(() => new Error(`Error while unlocking the case`))
+      ;(updateLockStatusToUnlocked as jest.Mock).mockImplementationOnce(
+        () => new Error(`Error while unlocking the case`)
+      )
 
       let result
       try {
